app/Http/Requests/Student/Attendanceディレクトリが肥大化しているための整理(yuki)
