+++ app/Http/Requests/Student/Attendance/ProgressRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Student;
+namespace App\Http\Requests\Student\Attendance;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class AttendanceIndexRequest extends FormRequest
+class ProgressRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -16,6 +16,13 @@ class AttendanceIndexRequest extends FormRequest
         return true;
     }
 
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'attendance_id' => $this->route('attendance_id'),
+        ]);
+    }
+
     /**
      * Get the validation rules that apply to the request.
      *
@@ -24,7 +31,7 @@ class AttendanceIndexRequest extends FormRequest
     public function rules()
     {
         return [
-            'search_word' => 'string',
+            'attendance_id' => ['required', 'integer', 'exists:attendances,id'],
         ];
     }
 }
