+++ app/Http/Requests/Student/Attendance/ShowChapterRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Student;
+namespace App\Http\Requests\Student\Attendance;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class AttendanceShowRequest extends FormRequest
+class ShowChapterRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -19,7 +19,9 @@ class AttendanceShowRequest extends FormRequest
     protected function prepareForValidation()
     {
         $this->merge([
+            'course_id' => $this->route('course_id'),
             'attendance_id' => $this->route('attendance_id'),
+            'chapter_id' => $this->route('chapter_id'),
         ]);
     }
 
@@ -32,6 +34,8 @@ class AttendanceShowRequest extends FormRequest
     {
         return [
             'attendance_id' => ['required', 'integer', 'exists:attendances,id'],
+            'course_id' => ['required', 'integer', 'exists:courses,id'],
+            'chapter_id' => ['required', 'integer', 'exists:chapters,id'],
         ];
     }
 }
