+++ app/Http/Requests/Student/Attendance/ShowRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Student;
+namespace App\Http\Requests\Student\Attendance;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class AttendanceShowChapterRequest extends FormRequest
+class ShowRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -19,9 +19,7 @@ class AttendanceShowChapterRequest extends FormRequest
     protected function prepareForValidation()
     {
         $this->merge([
-            'course_id' => $this->route('course_id'),
             'attendance_id' => $this->route('attendance_id'),
-            'chapter_id' => $this->route('chapter_id'),
         ]);
     }
 
@@ -34,8 +32,6 @@ class AttendanceShowChapterRequest extends FormRequest
     {
         return [
             'attendance_id' => ['required', 'integer', 'exists:attendances,id'],
-            'course_id' => ['required', 'integer', 'exists:courses,id'],
-            'chapter_id' => ['required', 'integer', 'exists:chapters,id'],
         ];
     }
 }
